Fix the `graphie.label` shim to not modify arguments
Summary:
The old code modified the `typesetAsMath` variable, which
changed the `arguments` array, so when `oldLabel.apply(this, arguments)`
was called, it had the wrong arguments. Also, I handled the last
argument being an object (which means it is a style, and not a
`typesetAsMath` flag) incorrectly.

Test Plan:
 - Ensure that this produces a red "x" label:
```
init({ range: [[0, 2], [0, 2]], scale: [40,40] });
label([1, 1], "x", "center", {color: "red"});
```
 - Ensure that the style of the label is saved when creating an
   svg+labels.

Reviewers: eater

Reviewed By: eater

Differential Revision: https://phabricator.khanacademy.org/D14916