Add timeout to phantomjs calls
Summary:
When a gunicorn worker thread is killed by the 30 second worker
timeout, any subprocesses it had created (such as phantomjs runners) are
not killed, and are left orphaned, which leads to memory problems for
the machine. This adds a timeout function to the subprocess calls to
ensure that any phantomjs processes are killed before the gunicorn
timeout occurs.

Test Plan:
 - Edit phantom_svg.js and comment out the `phantom.exit()` line, to
   simulate a long-running phantomjs runner
 - Make a request to render an image
 - Check `ps aux | grep phantomjs`, see that a phantomjs runner is
   running
 - Wait 25 seconds
 - See that the response returns a 500 error.
 - Check `ps aux | grep phantomjs`, see that there are no more phantomjs
   runners

Reviewers: eater

Reviewed By: eater

Differential Revision: https://phabricator.khanacademy.org/D15703