Make the timeout actually kill the phantomjs runners
Summary: We correctly called `kill` when the timeout was up, but this
only killed one of the 2 processes that were created by the
subprocess.Popen. This creates a new process group for the subprocesses
and then kills the whole process group at the end. This should fix the
problem of having a bunch of phantomjs runners running.

Test plan:
 - Run the server
 - run `curl --data "js=init(%7Brange%3A%20%5B%5B-1%2C%201%5D%2C%20%5B-1%2C%201%5D%5D%2Cscale%3A%20100%7D)%3Bfor%20(var%20i%20%3D%200%3B%20i%20%3C%201000%3B%20i%2B%2B)%20image(%22mouse%22%2C%20%5B0%2C%200%5D%2C%201)%3B" "http://localhost:8765/svg"`
 - Wait for it to return a timeout message
 - Quickly run `ps aux | grep phantomjs`, and see that there is nothing there.

Auditors: eater, csilvers