Always stringify content in labels.
Summary: Previously, we didn't stringify the content all the
time (sometimes they were numbers), which was fine (but kinda annoying
for external parsing). However, when we filter based on
`content.length`, the filter fails and we don't include numeric labels.

Instead of just stringifying when we test in the filter, just stringify
all the time.

Test plan:
 - Render
```
init({
    range: [[0, 2], [0, 2]]
    });

label([1, 1], 2);
```
 - See that `getData().labels` has one element in it.
 - See that `getData().labels[0].content` is a string, not a number.

Auditors: eater