Substate khan-exercises to include new colors in graphie-to-png
Summary: In D18158, I added new colors to khan-exercises from the new branding guide. These are now available to Perseus and Khan Exercises, but graphie-to-png doesn't yet have access to these new colors because it's using an older version of khan-exercises. So I simply substated khan-exercises to the newest version, which includes the new colors.

Test Plan:
- Start running graphie-to-png
- Load up sample snippet "1. Simple plot" (it's in the dropdown next to the Regraph button)
- On line 20, change `BLUE` to `GOLD_C` or another new color.
- Hit Regraph and ensure that the parabola turns the color you specified.

Reviewers: alex

Reviewed By: alex

Subscribers: cameron

Differential Revision: https://phabricator.khanacademy.org/D18441