Roll back Khan Exercises and add warnings.
Summary:
In January, some changes were made to Khan Exercises that breaks
an assumption that graphie-to-png makes. Namely that graphie will
always render everything into a single SVG.

This caused a regression that affected a number of content creators.
See https://app.asana.com/0/27216215224639/36842953088193 for the
support ticket.

This change brings us back to stable behavior in gaphie-to-png. But
now we're stuck. If we ever want to upgrade graphie-to-png to be up
to date with Khan Exercises again, we need to do a bunch of work. I've
added some warnings about this in the README that will hopefully
prevent us from upgrading Khan Exercises without considering this.

I've cherry picked in the color changes into a new graphie-to-png
branch of Khan Exercises to prevent regressions from that.

Test Plan:
A.1) Serve graphie-to-png locally
A.2) Go to http://localhost:5001/#27
A.3) See beautiful graph, and no errors

B.1) Serve graphie-to-png locally
B.2) Go to http://localhost:5001/#1
B.3) Change all references to `BLUE` to `GOLD_C`
B.4) Regraph
B.5) See the line turn gold

I cannot get "Convert to Image" to work locally, despite my best
efforts. I have confirmed that everything is being rendered into a
single SVG, which gives me high confidence. I will also test once this
is deployed to AWS (and rolling back is trivial), so I think this is
sufficient.

Reviewers: aria, emily

Reviewed By: emily

Differential Revision: https://phabricator.khanacademy.org/D18542