Make graphie-to-png able to handle web+graphie links!
Summary:
One of the benefits of using web+graphie links (theoretically) is that
we can register a handler for them so that when you click on them they
automatically load in graphie-to-png. Unfortunately, Hipchat didn't linkify
web+graphie links correctly (it would do web+[graphie://](...)), so it didn't
have a lot of use. But now, slack does linkify things correctly! So this is
useful!

This adds that handler, and adds some support for loading an image from a
`preload=` query parameter.

Test Plan:
 - Change `graphie-to-png.kasandbox.org` in the code to `localhost:8765`
 - Visit http://localhost:8765/
 - Click the content handler button in the chrome menu bar to allow localhost
   to handle web+graphie links.
 - Click on a web+graphie link from slack.
 - See that it opens http://localhost:8765/?preload=web+graphie://....
 - See that the image loads.

Reviewers: eater

Reviewed By: eater

Differential Revision: https://phabricator.khanacademy.org/D23529