[cp fixit] Allow "transform-origin" style in labels
Summary: In graphie-to-png, we have a list of allowed styles that content
creators can add to their labels. We don't allow all styles because some might
not work the same in all browsers/for all users.

This allows the "transform-origin" style, which we

Asana task:
https://app.asana.com/0/253524899192327/236022954302434

Test Plan:
 - `pip install -r requirements.txt`
 - `./server.sh`
 - Visit http://localhost:8765/
 - Enter the following program:
```
init({
    range: [[-3, 3], [-3, 3]],
    scale: 20
});
label([0, 0], "foo")
    .css("transform", "rotate(90deg)")
    .css("transform-origin", "left");
label([0, 0], "bar")
    .css("transform", "rotate(90deg)");
```
 - Click "regraph"
 - See that the labels successfully graph and that they are not
   overlapping (because the transform origin was successfully applied).

Auditors: aasmund