Changed setup to force equal scales on axes
The previous version defaulted to a square grid such that changes to the range resulted in different scales on the x- and y-axis. This made the vectors appear to have differing magnitudes (even though they really all had the same magnitude) as the slope changed. The use-nonsquare option makes the grid default to using the same scale for both axes, although there is a boolean at the top to revert it to the previous version if desired.