Use a specific version of Flask
Summary: When we recently tried to set up a new internal-services instance,
graphie-to-png failed to start because some dependencies of Flask didn't get
installed. It's unclear why they didn't get installed, but it seems that using
a specific version of Flask fixes this problem. (yay dependencies)

Test Plan:
 - `pip uninstall -r Flask itsdangerous click`
 - `pip install Flask==0.10.1`
 - See that graphie-to-png successfully starts

Auditors: colin