FIX: Do not block backspace in ACE search field
We have a code that blocks backspace key to
prevent accidental webpage closing
and another code that blocks that code in input fields.

This approach did not work for ACE search field because:
a) the search field element is added dynamically
when Ctrl+F is first pressed
b) it does not have the "type" attribute

Test Plan:
  - Run `python app.py`
  - Open `http://localhost:5001/`
  - Try typing in 'Load rendered image' input field. Backspace should
    work as expected.
  - Open the ACE search field by Ctrl+F and test that backspace works.
  - Focus out of the field, backspace should trigger a "Are you sure you
want to leave this page" alert window.