Update @khanacademy/mathjax-renderer to 2.1.0 (#24)
The new version just adds some macros for international trig operators.  This
is unlikely to be relevant for graphie-to-png, but I'm upgrading to be on the
safe side.

Issue: LEMS-2254

## Test plan:
Edit and export a graphie. Paste the URL into an image widget in the exercise
editor, and you should see the graphie with your changes.

Author: benchristel

Reviewers: mpolyak

Required Reviewers:

Approved By: mpolyak

Checks:

Pull Request URL: https://github.com/Khan/graphie-to-png/pull/24